+++ elfo/benches/round_trip.rs
@@ -113,8 +113,7 @@ async fn run(producer_count: u32, consumer_count: u32, mode: Mode, iter_count: u
     configurers.mount(elfo::configurer::fixture(&topology, AnyConfig::default()));
 
     elfo::_priv::do_start(topology, |ctx, _| async move {
-        ctx.request(Summarize)
-            .from(producers_addr)
+        ctx.request_to(producers_addr, Summarize)
             .all()
             .resolve()
             .await